+++ src/components/ColorPreview/index.ts
@@ -1,0 +1,1 @@
+export {ColorPreview, ColorPreviewType} from './ColorPrewiew';