+++ src/components/PieChart/components/Legend/components/index.ts
@@ -1,0 +1,1 @@
+export {LegendPreview} from './LegendPreview';