+++ src/components/PieChart/components/index.ts
@@ -1,0 +1,2 @@
+export {Arc} from './Arc';
+export {Legend} from './Legend';