+++ src/components/PieChart/types.ts
@@ -1,0 +1,7 @@
+import {Color} from '../../types';
+
+export interface Series {
+  name: string;
+  color: Color;
+  value: number;
+}